Move global configuration to Appearance category
