+++ README.adoc
@@ -27,6 +27,7 @@ Release notes are recorded in https://github.com/jenkinsci/custom-folder-icon-pl
 
 This version requires Jenkins 2.357 and above in order to support the transition to https://www.jenkins.io/blog/2022/06/28/require-java-11/[Java 11].
 
+* Version 2.11 moved the <<Global Configuration>> to the _Appearance_ configuration
 * Version 2.10 enables users to select the jobs to be considered for the combined build status in `BuildStatusFolderIcon`.
 
 .Changes in previous versions ...
