+++ src/main/resources/index.jelly
@@ -24,6 +24,6 @@ SOFTWARE.
 
 <?jelly escape-by-default='true'?>
 <div>
-    This plugin extends the <a href="https://github.com/jenkinsci/cloudbees-folder-plugin" target="_blank">Folders Plugin</a> to provide custom icons for folders.
+    This plugin extends the <a href="https://github.com/jenkinsci/cloudbees-folder-plugin" target="_blank">Folders Plugin</a> to provide custom icons for folders.<br/>
     You can upload your own images, use predefined icons and emojis or the combined build status of the jobs within a folder as icon.
 </div>
