+++ src/main/resources/jenkins/plugins/foldericon/BuildStatusFolderIcon/help-selection.html
@@ -23,5 +23,6 @@ SOFTWARE.
 -->
 
 <div>
-    You can select the jobs to be considered for the combined build status.</br>If you select none then all jobs are considered.
+    You can select the jobs to be considered for the combined build status.<br/>
+    If you select none then all jobs are considered.
 </div>
