+++ src/main/resources/jenkins/plugins/foldericon/CustomFolderIcon/help-upload.html
@@ -23,12 +23,12 @@ SOFTWARE.
 -->
 
 <div>
-    Upload an image to represent the folder.</br>
-    Use the "Browse..." button to select an image from disk.</br>
-    You can crop the image to the desired result and upload it using the "Apply" button.</br>
-    The file name will be randomized during upload.</br>
-    You can also select an image from the list of the already available icons.</br>
-    </br>
-    The recommended minimum size of the image is 128x128 pixels.</br>
+    Upload an image to represent the folder.<br/>
+    Use the "Browse..." button to select an image from disk.<br/>
+    You can crop the image to the desired result and upload it using the "Apply" button.<br/>
+    The file name will be randomized during upload.<br/>
+    You can also select an image from the list of the already available icons.<br/>
+    <br/>
+    The recommended minimum size of the image is 128x128 pixels.<br/>
     The image will be deleted automatically if this folder is being deleted - unless of course it is still used by another folder.
 </div>
