+++ src/main/resources/jenkins/plugins/foldericon/CustomFolderIconConfiguration/global.properties
@@ -19,9 +19,10 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
 CustomFolderIcons=Custom Folder Icons
-CleanupIcons=Cleanup Icons
-Cleanup=Cleanup
+AvailableIcons=Show all available icons
+CleanupIcons=Cleanup unused icons
+IconCount=Number of icons:
+DiskUsage=Disk usage of icons:
 CleanupSuccess=Successfully removed all unused icon images
 CleanupFailed=Failed to removed unused icon images:
