+++ src/main/resources/jenkins/plugins/foldericon/CustomFolderIconConfiguration/help-cleanup.html
