+++ src/main/resources/jenkins/plugins/foldericon/EmojiFolderIcon/config.jelly
@@ -31,7 +31,7 @@ SOFTWARE.
         <l:icon id="emoji-preview" src="${emojis.get(instance.emoji ?: 'sloth')}" title="${icon.key}"
                 class="icon-xlg" />
     </f:entry>
-    <f:advanced title="${%AvailableIcons}">
+    <f:advanced title="${%AvailableIcons}${not empty emojis ? ' (' + emojis.size() + ')' : ''}">
         <j:forEach var="icon" items="${emojis}">
             <a id="select-emoji-${icon.key}" class="emoji-icon-selection" onclick="setEmojiIcon('${icon.key}')">
                 <l:icon id="emoji-icon-${icon.key}" src="${icon.value}" title="${icon.key}" tooltip="${icon.key}"
