+++ src/main/webapp/js/custom-icon-global.js
@@ -29,15 +29,15 @@
  * @param {string} errorMessage - The error message.
  */
 function doCustomIconCleanup(successMessage, errorMessage) {
-    fetch(rootURL + "/descriptor/jenkins.plugins.foldericon.CustomFolderIcon/cleanup", {
+    fetch(rootURL + "/descriptor/jenkins.plugins.foldericon.CustomFolderIconConfiguration/cleanup", {
         method: "post",
         headers: crumb.wrap({}),
     }).then(rsp => {
         let button = document.getElementById("custom-icon-cleanup")
         if (rsp.ok) {
-            hoverNotification(successMessage, button.parentNode);
+            hoverNotification(successMessage, button.parentNode.parentNode);
         } else {
-            hoverNotification(errorMessage + " " + rsp.status + " - " + rsp.statusText, button.parentNode);
+            hoverNotification(errorMessage + " " + rsp.status + " - " + rsp.statusText, button.parentNode.parentNode);
         }
     }).catch(error => {
         console.error(error);
